+++ CHANGELOG.md
@@ -484,6 +484,11 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Recursion
 
+## 0103 - Binary Tree Zigzag Level Order Traversal
+
+- Recursion
+- Orientation
+
 ## 0415 - Add strings
 
 - Full adder